+++ test/td_df_lib/parser_test.exs
@@ -277,6 +277,46 @@ defmodule TdDfLib.ParserTest do
                  domain_ids: []
                })
     end
+
+    test "foo" do
+      schema = [
+        %{
+          "cardinality" => "?",
+          "default" => %{"value" => "", "origin" => "default"},
+          "description" => "Input your integer",
+          "label" => "Number",
+          "name" => "input_integer",
+          "type" => "integer",
+          "widget" => "number",
+          "values" => nil
+        },
+        %{
+          "cardinality" => "?",
+          "default" => %{"value" => "", "origin" => "default"},
+          "description" => "Input your float",
+          "label" => "Number",
+          "name" => "input_float",
+          "type" => "float",
+          "widget" => "number",
+          "values" => nil
+        }
+      ]
+
+      content = %{
+        "input_float" => %{"origin" => "file", "value" => "foo"},
+        "input_integer" => %{"origin" => "file", "value" => "bar"}
+      }
+
+      fields =
+        Parser.format_content(%{
+          content: content,
+          content_schema: schema,
+          domain_ids: []
+        })
+
+      assert fields["input_float"]["value"] == {:error, :invalid_format}
+      assert fields["input_integer"]["value"] == {:error, :invalid_format}
+    end
   end
 
   describe "get_from_content/2" do